Multi-Agent-Talent-Acquistion/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
PyMuPDF==1.24.5
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.0
faiss-cpu==1.7.2
langgraph==0.4.5
langgraph-cli==0.2.10
langsmith==0.1.129
python-docx==1.1.0
structlog==24.1.0
streamlit==1.32.0
openai==1.44.1
pymysql==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
sqlalchemy==2.0.29
tenacity==8.2.3
httpx==0.27.0
pandas==2.2.2
PyMySQL== 1.1.1
Levenshtein==0.27.1 
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1 
rapidfuzz==3.13.0